Who Killed the (End-User) Written Language?
First(ish) there was the shell, this append-only log of interactions
where you talk to the computer by typing instructions and receive
responses.

(My history may be off by a lot here, but bear with me and assume
terminals have been as they are today more or less unchanged. The
relevant version of history is sitting right in front of me while I type
these lines. As far as I know, the most modern, GPU-rendered,
web-enabled, etc. terminal emulators of today are still, as the name
implies, _emulating_ computer systems that were alive in the 70's.)

Then came GUIs, the desktop metaphor, Windows and all that, and that too
hasn't changed substantially. You drag rectangles around with a mouse,
resize them, arrange them next or on top of each other, with partial
overlap in many cases. Except for the specific action of typing into
text fields, the dominant mode of interaction for most people is
clicking on buttons and widgets using a mouse, touchpad, or touch
screen.

There is an in-between that we call TUIs, which are a way to display
window-like things that take up the screen and can present interactive
elements spatially. They can be navigated with the keyboard or mouse,
and things can also be clicked on either with the actual mouse or by
moving focus with the keyboard and hitting a key like Enter.

Sometimes the clicked-on thing is invisible, in which case we call it
"keyboard shortcut", "keybinding", "hotkey", "keymap", etc. (It seems
like there has been enough time to settle on naming there, but the they
are still legion for some reason.) They are categorically related to GUI
widgets in the sense that they don't, in principle, produce a
user-visible data representation of the action taken that is then added
to the end of a history as in the case of the shell.

And here is the crucial thing that was lost in the transition: History.

The shell can be viewed as a 1D conversational metaphor where the
one and dominant dimension is time. You can scroll up and see what
you've told the computer and what it answered back going as far back as
your memory and settings allow. You can search through it. You can reuse
previously-typed commands verbatim, edit, and remix them.

The desktop GUI flattened all that into a 2D snapshot of current state.
No going back except for some limited, app-specific undo/redo action. No
introspection, no reuse.

As I read Paulo Freire's work on natural-language literacy and the
life-changing power of externalizable cognition, I started seeing very
close parallels, and was filled with grief.

We lost *written language* itself and could no longer talk about our
actions in precise ways, or represent them outside our heads to think
about them, manipulate them, and build upon them. The sadness of
instruction manuals filled with screenshots points to the tragedy of
obliterated literacy. We went back to representational cave paintings.

By all accounts, the desktop metaphor came out of Xerox PARC and it's
still cited as this major leap in user-friendliness and interactivity.
I'm of the opinion that it set the computing world on a wrong trajectory
that cost us decades of time, and immesurable loss of cognitive
potential.

And yet, we have a hard time seeing how it could be different. Due to
historic coupling, it just seem that advanced graphics and the desktop
metaphor are pretty much one and the same. How would you even
_conceive_ of such an odd thing as a graphical shell? What would that
anachronistic contraption even _look_ like? How would the untrained user
even know what to do?

Those are not rhetorical questions ;)